requirements.txt
youtube-dl
ffmpeg-python
requests
icecream
mysql-connector-python
beautifulsoup4
sentry_sdk
git+https://github.com/openai/whisper.git
python-dotenv
yt-dlp
python-docx
polling